obexd: Allow plugins to create Unicode headers
This is a temporary solution that should be written more cleanly.

CTN profile specifies that in response to a PUT used to upload an
object the server must include Name header containing the handle
for the newly created object.

Allow get_next_header to construct headers with Unicode values
to achieve this.