gobex: Expose final bit to data consumers
A profile implementation may need to do a potentially expensive
operation such as committing received PUT data into storage or
validating it syntactically when the last chunk of the request is
received. Passing the "final" bit as an argument to GObexDataConsumer
functions allows the function to see when end of data is reached.